Reworked tests as typescript
Most tests are now redundant due to static type checking.
