+++ package.json
@@ -9,7 +9,7 @@
   "browser": "index.js",
   "scripts": {
     "start": "webpack serve --mode development",
-    "test": "nyc mocha -r ts-node/register --recursive test",
+    "test": "TS_NODE_PROJECT=\"./test/tsconfig.json\" nyc mocha -r ts-node/register test/**/*.test.ts",
     "lint:scripts": "eslint --ext .js,.jsx,.ts,.tsx src/ test/",
     "lint:styles": "stylelint src/**/*.scss",
     "clean": "rm -rf dist/",
@@ -32,8 +32,10 @@
     "@babel/runtime-corejs3": "^7.23.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@reduxjs/toolkit": "^2.0.1",
+    "@types/chai": "^4.3.11",
     "@types/d3": "^7.4.3",
     "@types/draft-js": "^0.11.16",
+    "@types/mocha": "^10.0.6",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@types/react-redux": "^7.1.32",
