+++ src/common/scripts/partnerTypes.ts
