+++ test/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es6",
+    "lib": [
+      "ES2021.WeakRef",
+      "dom"
+    ],
+    "plugins": [{ "name": "typescript-plugin-css-modules" }],
+    "sourceMap": true,
+    "esModuleInterop": true,
+    "jsx": "react",
+    "types": ["mocha", "typescript", "node"],
+  },
+  "exclude": [
+    "node_modules"
+  ]
+}
